[Silver V] Title: 최소공배수, Time: 108 ms, Memory: 14100 KB -BaekjoonHub

+++ 백준/Silver/13241. 최소공배수/최소공배수.java
@@ -4,8 +4,8 @@ import java.io.*;
 
 class Main {
     
-    static int readPosInt() throws IOException{
-        int c, n = System.in.read() & 15;
+    static long readPosInt() throws IOException{
+        long c, n = System.in.read() & 15;
         while((c = System.in.read()) > 32) n = (n << 3) + (n << 1) + (c & 15);
         return n;
     }
@@ -15,9 +15,9 @@ class Main {
         int exp = 0;
         
         while(((a|b)&1) == 0){
-                exp++;
-                a >>= 1;
-                b >>= 1;
+            exp++;
+            a >>= 1;
+            b >>= 1;
         }
             
         while(a != 0 && b != 0){
@@ -29,7 +29,7 @@ class Main {
 
         if(a > b) return(axb>>exp)/a;
         else return(axb>>exp)/b;
-        
+     
     }
     
     public static void main(String[] args) throws IOException{
